First version 1.0.0 of the TelaMod by Anthonious333 for Minecraft ver 1.20.4
Added: TelaRail, DestaRail, and Tela Wrench

TelaRail: When provided with ender pearls and a redstone signal it will teleport any mine cart on it to the conected DestaRail.

DestaRail: The destination of the TelaRail. Any number of TelaRails can connect to one DestaRail.

Tela Wrench: Used to connect the DestaRail and the TelaRail. First, right click on the DestaRail you want to be the destination. Then, shift right click on the TelaRail you would like to lead to your DestaRail. Please note that once you select a DestaRail, all Tela Wrenches will lead to that block until the world is closed or a new destination is selected.